Add an MVP for time-split
Adds a first mvp for the timesplit interface. Uses a hard coded list of
acceptable checkpoints (e.g. when a specific research is finished).
Allows the user to specify at what timestamp each checkpoint is reached
in the mod settings page. Will display the previous time, current time
and a diff for each item reached. Includes a diff for the current
un-reached item as well. The implementation is otherwise rough and lacks
the ability to save a run and automatically use it for the next game
(i.e. user has to explicitly set the timestamps - and we do not provide
an easy way to do so)

Does not seem to be an easy way to save parameter values in one game to
be used in a new game (the global instance is a per-save only)